+++ provider-loopme/src/main/java/com/deltadna/android/sdk/ads/provider/loopme/LoopMeAdapter.java
@@ -63,12 +63,17 @@ public final class LoopMeAdapter extends MediationAdapter {
             MediationListener listener,
             JSONObject configuration) {
         
-        ad = LoopMeInterstitial.getInstance(appKey, activity);
-        if (ad != null) {
-            this.listener = listener;
-            forwarder = new EventForwarder(listener, this);
+        this.listener = listener;
+        
+        if (ad == null) {
+            forwarder = new EventForwarder(this).setListener(listener);
             
+            ad = LoopMeInterstitial.getInstance(appKey, activity);
+            ad.setAutoLoading(false);
             ad.setListener(forwarder);
+            ad.load();
+        } else if (forwarder != null) {
+            forwarder.setListener(listener);
             ad.load();
         }
     }
@@ -82,7 +87,10 @@ public final class LoopMeAdapter extends MediationAdapter {
             } else if (!ad.isReady() && listener != null) {
                 Log.w(BuildConfig.LOG_TAG, "Ad is not ready");
                 listener.onAdFailedToShow(this, AdShowResult.EXPIRED);
-            } else if (ad.isReady()) {
+            } else if (!ad.isReady() && listener != null) {
+                Log.w(BuildConfig.LOG_TAG, "Ad is not loaded");
+                listener.onAdFailedToShow(this, AdShowResult.NOT_LOADED);
+            } else {
                 ad.show();
             }
         }
@@ -114,4 +122,9 @@ public final class LoopMeAdapter extends MediationAdapter {
     
     @Override
     public void onResume() {}
+    
+    @Override
+    public boolean isGdprCompliant() {
+        return true;
+    }
 }